Moves extension methods to the Sustainsys.Saml2.AspNetCore namespace to align with project structure and Removed all unused using statements.
