+++ src/Sustainsys.Saml2/AspNetCore/Events/AuthnRequestGeneratedContext.cs
@@ -1,6 +1,5 @@
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Http;
-using Sustainsys.Saml2.AspNetCore;
 using Sustainsys.Saml2.Samlp;
 
 namespace Sustainsys.Saml2.AspNetCore.Events;
