+++ src/Sustainsys.Saml2/Bindings/HttpPostBinding.cs
@@ -1,9 +1,5 @@
 ﻿using Microsoft.AspNetCore.Http;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 using System.Xml;
 
 namespace Sustainsys.Saml2.Bindings;
