+++ src/Sustainsys.Saml2/Metadata/SSODescriptor.cs
@@ -1,6 +1,4 @@
-﻿using Sustainsys.Saml2.Xml;
-
-namespace Sustainsys.Saml2.Metadata;
+﻿namespace Sustainsys.Saml2.Metadata;
 
 
 /// <summary>
