+++ src/Sustainsys.Saml2/Saml/Assertion.cs
@@ -1,10 +1,4 @@
 ﻿using Sustainsys.Saml2.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Sustainsys.Saml2.Saml;
