+++ src/Sustainsys.Saml2/Saml/Subject.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sustainsys.Saml2.Saml;
+﻿namespace Sustainsys.Saml2.Saml;
 
 /// <summary>
 /// A Saml2 Subject, see core 2.4.1.
