+++ src/Sustainsys.Saml2/Saml2Entity.cs
@@ -1,9 +1,4 @@
 ﻿using Sustainsys.Saml2.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Sustainsys.Saml2;
 /// <summary>
