+++ src/Sustainsys.Saml2/Serialization/ISamlXmlWriter.cs
@@ -1,5 +1,4 @@
-﻿using Sustainsys.Saml2.Saml;
-using Sustainsys.Saml2.Samlp;
+﻿using Sustainsys.Saml2.Samlp;
 using System.Xml;
 
 namespace Sustainsys.Saml2.Serialization;
