+++ src/Sustainsys.Saml2/Serialization/SamlXmlReader.AuthnContext.cs
@@ -1,7 +1,5 @@
-﻿using Sustainsys.Saml2.Metadata;
-using Sustainsys.Saml2.Saml;
+﻿using Sustainsys.Saml2.Saml;
 using Sustainsys.Saml2.Xml;
-using System.Diagnostics;
 using static Sustainsys.Saml2.Constants;
 
 namespace Sustainsys.Saml2.Serialization;
