+++ src/Sustainsys.Saml2/Serialization/SamlXmlReader.AuthnRequest.cs
@@ -1,6 +1,5 @@
 ﻿using Sustainsys.Saml2.Samlp;
 using Sustainsys.Saml2.Xml;
-using System.Xml;
 using static Sustainsys.Saml2.Constants;
 
 namespace Sustainsys.Saml2.Serialization;
