+++ src/Sustainsys.Saml2/Serialization/SamlXmlReader.RequestAbstractType.cs
@@ -1,11 +1,5 @@
-﻿using Microsoft.AspNetCore.Antiforgery;
-using Sustainsys.Saml2.Samlp;
+﻿using Sustainsys.Saml2.Samlp;
 using Sustainsys.Saml2.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using static Sustainsys.Saml2.Constants;
 
 namespace Sustainsys.Saml2.Serialization;
