+++ src/Sustainsys.Saml2/Validation/SamlValidationException.cs
@@ -1,11 +1,4 @@
-﻿using Sustainsys.Saml2.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sustainsys.Saml2.Validation;
+﻿namespace Sustainsys.Saml2.Validation;
 
 /// <summary>
 /// Exception thrown when validation is not successful
