+++ src/Sustainsys.Saml2/Xml/ReadErrorInspectorContext.cs
@@ -1,10 +1,4 @@
-﻿using Microsoft.Extensions.Configuration.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml;
+﻿using System.Xml;
 
 namespace Sustainsys.Saml2.Xml;
