+++ src/Sustainsys.Saml2/Xml/XmlTraverser.cs
@@ -1,7 +1,4 @@
 ﻿using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
-using System.Runtime.CompilerServices;
-using System.Runtime.ExceptionServices;
 using System.Security.Cryptography.Xml;
 using System.Xml;
