+++ src/Tests/Sustainsys.Saml2.Tests/AspNetCore/Saml2HandlerTests.cs
@@ -5,11 +5,8 @@ using Microsoft.Extensions.Options;
 using NSubstitute;
 using Sustainsys.Saml2.Bindings;
 using Sustainsys.Saml2.Xml;
-using Sustainsys.Saml2.Saml;
 using Sustainsys.Saml2.Samlp;
-using System.Net.Http;
 using System.Text.Encodings.Web;
-using System.Xml;
 using Sustainsys.Saml2.Tests.Helpers;
 using System.Text;
 using Sustainsys.Saml2.Serialization;
