+++ src/Tests/Sustainsys.Saml2.Tests/Bindings/HttpPostBindingTests.cs
@@ -1,12 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http;
 using NSubstitute;
 using Sustainsys.Saml2.Bindings;
-using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Xml;
 
 namespace Sustainsys.Saml2.Tests.Bindings;
