+++ src/Tests/Sustainsys.Saml2.Tests/Helpers/XmlExtensions.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml;
+﻿using System.Xml;
 
 namespace Sustainsys.Saml2.Tests.Helpers;
 public static class XmlExtensions
