+++ src/Tests/Sustainsys.Saml2.Tests/Saml/NameIdTests.cs
@@ -1,9 +1,4 @@
 ﻿using Sustainsys.Saml2.Saml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Sustainsys.Saml2.Tests.Saml;
 public class NameIdTests
