+++ src/Tests/Sustainsys.Saml2.Tests/Serialization/SamlXmlReaderTests.Assertion.cs
@@ -1,15 +1,7 @@
-﻿using FluentAssertions.Equivalency.Steps;
-using Microsoft.Extensions.DependencyInjection;
-using Sustainsys.Saml2.Saml;
+﻿using Sustainsys.Saml2.Saml;
 using Sustainsys.Saml2.Serialization;
 using Sustainsys.Saml2.Tests.Helpers;
 using Sustainsys.Saml2.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Text;
-using System.Threading.Tasks;
 using System.Xml;
 
 namespace Sustainsys.Saml2.Tests.Serialization;
