+++ src/Tests/Sustainsys.Saml2.Tests/Xml/XmlHelperTests.cs
@@ -1,10 +1,4 @@
 ﻿using Sustainsys.Saml2.Xml;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
-using System.Threading.Tasks;
 using System.Xml;
 
 namespace Sustainsys.Saml2.Tests.Xml;
